Update get_ir() to include X,head_orientation,xs
Now we calculate the desired directon directly in get_ir(), which means we have
to change ir_generic() accordingly.
We also introduce two cases in get_ir() depening on if we are using HRIR or
BRIR. For the last one, now head orientation is independently considered from
source position and the sofa file can include serveral loudsepakers now.

The code for the BRIRs returns the wrong impulse response at the moment and
needs some fixing.