Add additional functon for calculating zeros of Hankel function
See https://github.com/sfstoolbox/sfs/issues/57#issuecomment-186192718
The result seems to be more or less identical to the one obtained with
sphbesselh_zeros() in terms of stability -- the obtained values for z, p differ
quite a lot.
One advantage of sphhankel_laplace() is that it will not lead to Inf or Nan, but
only to large numerical errors in the reproduced sound field for orders >85.

Maybe it would be a good idea to try to calculate z directly via an formula
without using the zeros() function.